WIP: local auth 1/3
