+++ backend/core/schemas/security/__init__.py
@@ -1,15 +1,14 @@
 # -*- coding: utf-8 -*-
 """Pydantic security models."""
 
-from .auth import BaseUserModel, UserCreateModel, UserViewModel
-from .oauth2 import AccessToken, GoogleIdInfo, RefreshToken, Token
+from .oauth2 import AccessToken, RefreshToken, Token, TokenData
+from .user import UserCreateModel, UserViewModel
 
 __all__ = [
-    "BaseUserModel",
     "UserCreateModel",
     "UserViewModel",
     "Token",
     "AccessToken",
     "RefreshToken",
-    "GoogleIdInfo",
+    "TokenData",
 ]
